% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.proba-package}
\alias{mlr3learners.proba}
\alias{mlr3learners.proba-package}
\title{mlr3learners.proba: Learners for mlr3proba.}
\description{
Learners for mlr3proba.
}
\author{
\strong{Maintainer}: Raphael Sonabend \email{raphael.sonabend.15@ucl.ac.uk} (\href{https://orcid.org/0000-0001-9225-4654}{ORCID})

}
